+++ slack-gamebot/api/endpoints.rb
@@ -1,4 +1,5 @@
 require 'slack-gamebot/api/endpoints/challenges_endpoint'
 require 'slack-gamebot/api/endpoints/matches_endpoint'
 require 'slack-gamebot/api/endpoints/users_endpoint'
+require 'slack-gamebot/api/endpoints/seasons_endpoint'
 require 'slack-gamebot/api/endpoints/root_endpoint'