+++ slack-gamebot/api/presenters/user_rank_presenter.rb
@@ -1,21 +1,20 @@
 module Api
   module Presenters
-    module UserPresenter
+    module UserRankPresenter
       include Roar::JSON::HAL
       include Roar::Hypermedia
       include Grape::Roar::Representer
 
-      property :id, type: String, desc: 'User ID.'
-      property :user_name, type: String, desc: 'User name.'
+      property :id, type: String, desc: 'UserRank ID.'
+      property :user_name, type: String, desc: 'UserRank name.'
       property :wins, type: Integer, desc: 'Number of wins.'
       property :losses, type: Integer, desc: 'Number of losses.'
       property :elo, type: Integer, desc: 'Elo.'
       property :rank, type: Integer, desc: 'Rank.'
-      property :created_at, as: :registered_at, type: DateTime, desc: 'Date/time when the user has registered.'
 
-      link :self do |opts|
+      link :user do |opts|
         request = Grape::Request.new(opts[:env])
-        "#{request.base_url}/users/#{id}"
+        "#{request.base_url}/users/#{user_id}"
       end
     end
   end