+++ slack-gamebot/api/presenters/user_ranks_presenter.rb
@@ -1,0 +1,12 @@
+module Api
+  module Presenters
+    module UserRanksPresenter
+      include Roar::JSON::HAL
+      include Roar::Hypermedia
+      include Grape::Roar::Representer
+      include Api::Presenters::PaginatedPresenter
+
+      collection :results, extend: UserRankPresenter, as: :user_ranks, embedded: true
+    end
+  end
+end