+++ spec/fabricators/season_fabricator.rb
@@ -1,2 +1,5 @@
 Fabricator(:season) do
+  before_validation do
+    Fabricate(:match) if Challenge.current.none?
+  end
 end