+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.1.0 (Next)
 
+* [#4](https://github.com/dblock/slack-gamebot/issues/4): Added support for seasons - [@dblock](https://github.com/dblock).
 * [#17](https://github.com/dblock/slack-gamebot/pull/17): Show rank section for N players - [@wrgoldstein](https://github.com/wrgoldstein).
 * Initial public release - [@dblock](https://github.com/dblock).